+++ CHANGELOG.md
@@ -153,3 +153,4 @@
 * Added a hover indicator on WFS features 
 * Added fullscreen button
 * Updated dependencies and fixed new Ol10 types errors
+* Fixed LockId request